New IPMI functions
